pytest refactoring
